+++ Taller2_A00356210/src/main/java/com/example/model/controller/PersonController.java
@@ -47,12 +47,12 @@ public class PersonController {
 	}
 
 	@DeleteMapping("/{id}")
-	public void delete(@PathVariable("id") Long id) {
+	public void delete(@PathVariable("id") Integer id) {
 		personService.deleteById(id);
 	}
 
 	@GetMapping("/{id}")
-	public Person getById(@PathVariable("id") Long id) {
+	public Person getById(@PathVariable("id") Integer id) {
 		return personService.get(id).orElseThrow(() -> new IllegalArgumentException("Invalid id"));
 	}
