+++ Taller2_A00356210/src/main/java/com/example/services/DepartmentServiceIMPL.java
@@ -31,7 +31,7 @@ public class DepartmentServiceIMPL implements departmentService{
 		
 	
 	@Override
-	public Department findDepartmentById(long id) {
+	public Department findDepartmentById(Integer id) {
 		return dr.findById(id).get();
 	}
